Add settings to scroll on edition out of viewport
By default, when there is an edition of a line, which is out of the
viewport, Etherpad scrolls the minimum necessary to make this line
visible. This makes that the line stays either on the top or the bottom
of the viewport. With this commit, we add a setting to make possible to
scroll to a position x% pixels from the viewport. Besides of that, we
add
a setting to make an animation of this scroll.
If nothing is changed on settings.json the Etherpad default behavior is
kept.